[FEAT] Added header based on file name
When creating a brand new zettel entry, you're
prompted for the file name. My next step
would then be to go into the file and create
a header based on the file name. I take the name
variable and write it to file before I open
it in the editor for the user.